+++ game-server/src/main/java/com/sparky/trak/game/server/controller/ConsoleController.java
@@ -1,13 +1,13 @@
-package com.sparky.trak.game.webapp.controller;
+package com.sparky.trak.game.server.controller;
 
 import com.sparky.trak.game.repository.specification.ConsoleSpecification;
 import com.sparky.trak.game.service.ConsoleService;
 import com.sparky.trak.game.service.GameService;
 import com.sparky.trak.game.service.dto.ConsoleDto;
 import com.sparky.trak.game.service.dto.GameDto;
-import com.sparky.trak.game.webapp.assembler.ConsoleRepresentationModelAssembler;
-import com.sparky.trak.game.webapp.assembler.GameRepresentationModelAssembler;
-import com.sparky.trak.game.webapp.exception.ApiError;
+import com.sparky.trak.game.server.assembler.ConsoleRepresentationModelAssembler;
+import com.sparky.trak.game.server.assembler.GameRepresentationModelAssembler;
+import com.sparky.trak.game.server.exception.ApiError;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -41,7 +41,7 @@ public class ConsoleController {
      * its pre-requisite conditions in order to attempt a save to the persistence layer.
      *
      * If the {@link ConsoleDto} being saved contains an ID that matches an existing entity in the persistence layer,
-     * the {@link ConsoleDto} will not be saved and a {@link com.sparky.trak.game.webapp.exception.ApiError} will
+     * the {@link ConsoleDto} will not be saved and a {@link com.sparky.trak.game.server.exception.ApiError} will
      * be returned with appropriate exceptions details.
      *
      * @param consoleDto The {@link ConsoleDto} to save.
