+++ game-server/src/main/java/com/sparky/trak/game/server/controller/GameController.java
@@ -1,4 +1,4 @@
-package com.sparky.trak.game.webapp.controller;
+package com.sparky.trak.game.server.controller;
 
 import com.sparky.trak.game.repository.specification.GameSpecification;
 import com.sparky.trak.game.service.ConsoleService;
@@ -7,10 +7,10 @@ import com.sparky.trak.game.service.GenreService;
 import com.sparky.trak.game.service.dto.ConsoleDto;
 import com.sparky.trak.game.service.dto.GameDto;
 import com.sparky.trak.game.service.dto.GenreDto;
-import com.sparky.trak.game.webapp.assembler.ConsoleRepresentationModelAssembler;
-import com.sparky.trak.game.webapp.assembler.GameRepresentationModelAssembler;
-import com.sparky.trak.game.webapp.assembler.GenreRepresentationModelAssembler;
-import com.sparky.trak.game.webapp.exception.ApiError;
+import com.sparky.trak.game.server.assembler.ConsoleRepresentationModelAssembler;
+import com.sparky.trak.game.server.assembler.GameRepresentationModelAssembler;
+import com.sparky.trak.game.server.assembler.GenreRepresentationModelAssembler;
+import com.sparky.trak.game.server.exception.ApiError;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -57,7 +57,7 @@ public class GameController {
      * its pre-requisite conditions in order to attempt a save to the persistence layer.
      *
      * If the {@link GameDto} being saved contains an ID that matches an existing entity in the persistence layer,
-     * the {@link GameDto} will not be saved and a {@link com.sparky.trak.game.webapp.exception.ApiError} will
+     * the {@link GameDto} will not be saved and a {@link com.sparky.trak.game.server.exception.ApiError} will
      * be returned with appropriate exceptions details.
      *
      * @param gameDto The {@link GameDto} to save.
