+++ game-server/src/main/resources/application.yml
@@ -1,4 +1,10 @@
+eureka:
+  client:
+    service-url:
+      default-zone: http://localhost:8761/eureka
 spring:
+  application:
+    name: trak-game-server
   datasource:
     url: jdbc:h2:mem:game-db;DB_CLOSE_DELAY=-1
     driver-class-name: org.h2.Driver
