+++ game-server/src/test/java/com/sparky/trak/game/server/controller/GenreControllerIntegationTest.java
@@ -1,7 +1,7 @@
-package com.sparky.trak.game.webapp.controller;
+package com.sparky.trak.game.server.controller;
 
 import com.sparky.trak.game.service.dto.GenreDto;
-import com.sparky.trak.game.webapp.exception.ApiError;
+import com.sparky.trak.game.server.exception.ApiError;
 import org.apache.http.client.HttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.junit.jupiter.api.Assertions;
